Add simple UI for conversations
A list of conversations is shown in a sidebar on the left hand. The
messages occupy the main area on the right, using a React component (in
preparation for later dynamic adding of message). The messages are
positioned depending on whether they were sent or received by the
current player.