Add mix tasks for starting La Famiglia
Previously, the event loop would be started on every invocation of mix.
With this change, it has to be started explicitly, either by invoking
`mix la_famiglia.event_loop` which starts just the event loop or `mix
la_famiglia.server` which starts the event loop and the web server.

Note that the event loop is not automatically started on `mix
phoenix.server`. Actions that need to access the event queue will thus
fail with an ArgumentError.