Add priv/repo/seeds.exs
Currently, it uses a rather simple method of creating a small number of
players and villas. Due to the way `Forge.saved_player` works, the
database needs to be reset before the file can be run again. Unless the
database is reset the unique constraint is violated for the generated
email addresses.