Call `process_virtually_until` on event handling
Before this commit, it was possible that wrong numbers were displayed
for items in the unit queue. That was often due to `processed_until`
still having the value it had when the last item was enqueued. In that
case, `units_recruited_between` was fed with the wrong start time,
possibly returning a value greater than `item.number`.