Split `processed_until`
By keeping track of `resources_gained_until` and `units_recruited_until`
separately, it becomes much easier to reason about the state of a villa.
This is especially useful in event handling as the respective handlers
only have to deal with a minimal amount of changes which are independent
of each other.