Re-organize relation between reports and villas
Previously, there were two associations, `related_villas` on Report and
`related_reports` on Villa. Both these associations did not specify what
the relationship to the related records was like.

This commit introduces two associations on Report, `origin` and
`target`, which make the kind of relationship explicit and are therefore
much more useful beyond the initial scope of finding reports that were
related to a specific villa.