Refactor AttackDialog
Previously, AttackDialog was a component with a `main` function as well
as `Model` and `Msg` types for holding its internal state. Now,
AttackDialog is just a collection of stateless functions while all state
is held outside of it. The refactoring follows the example of
https://github.com/evancz/elm-sortable-table.

Custom handlers for the events originating from the dialog are passed
into `view` by using the respective fields of the type `Config`. What
happens when a user clicks on the buttons, is thus entirely up to the
caller of `view`. This makes the dialog easily reusable.