Run Webpack build on Travis
An Elm compile requires the module `Mechanics.Units` which is, as of
February 2020, only generated from an EEx template by a Mix task. To
keep both builds isolated, `assets/elm/Mechanics/Units.elm` is checked
into version control with this commit.

Remove dependency on phoenix_html

phoenix_html, in the previous setup, was expected to be located at
`../deps/phoenix_html`. That required `mix deps.get` to have been run
because it was configured as a local dependency, not loaded from npm.
Currently, we don’t want to setup Elixir when we run npm because that
would defeat the purpose of having a build matrix.