Send item to queue after commit
- Drop item from queue after commit
- Remove `Actions.Enqueue` and `Actions.Dequeue`
- Remove `Actions.Attack` and `Actions.Cancel`

Don’t use `Multi.run` because that runs inside the transaction, and
there is a risk the event handler tries to read the new item from the
database before it has been committed. Instead, send the item to the
queue (or drop it, respectively) inside the controller when we are sure
it has been committed.